fixed edit(we have been calling it update) function, commented a bit of code
It now works to add new users, delete new or hardcoded users, edit the name and color of hardcoded users. What we need is the ejs/html for showuser.ejs, the page that displays single user info. We also need a button on /users for each user where you can click to navigate to showuser/:id and it shows all their shit. Then clean up code a bit, comment on functions, update readme, then we done.